finished 4.23

+++ part4/bloglist/controllers/bloglist.js
@@ -2,8 +2,9 @@ const bloglistRouter = require('express').Router()
 const Blog = require('../models/bloglist')
 // by using express-async-errors, try...catch is not required
 
+
 bloglistRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
@@ -14,23 +15,39 @@ bloglistRouter.get('/:id', async (request, response) => {
 
 bloglistRouter.post('/', async (request, response) => {
     const body = request.body
+    const user = request.user
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
+        likes: body.likes,
+        user: user.id
     })
     const result = await blog.save()
+    user.blogs = user.blogs.concat(result.id)
+    await user.save()
+
     response.status(201).json(result)
+
 })
 
 bloglistRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const owner = await Blog.findById(request.params.id)
+    if(owner.user.toString() === request.user.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    }else{
+        response.status(401).json({ error: 'blog can only be deleted by creator' })
+    }
 })
 
 bloglistRouter.put('/:id', async (request, response) => {
-    await Blog.findByIdAndUpdate(request.params.id,{ likes: request.body.likes })
-    response.status(201).end()
+    const owner = await Blog.findById(request.params.id)
+    if(owner.user.toString() === request.user.id.toString()) {
+        await Blog.findByIdAndUpdate(request.params.id,{ likes: request.body.likes })
+        response.status(201).end()
+    }else{
+        response.status(401).json({ error: 'blog can only be deleted by creator' })
+    }
 })
 module.exports = bloglistRouter
